backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
edge-tts>=6.1.0
pyttsx3>=2.90
pygame>=2.5.0
pydantic>=2.6.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
faster-whisper>=1.0.0
